organzing code for better view
