--- src/components/AccountByAcctID.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import {AccountByAcctIDConst} from './AccountByAcctIDApi';
-
-function AccountByAcctID() {
-  return( 
-  <div>
-      <AccountByAcctIDConst/>
-  </div>
-  )
-}
-
-export default AccountByAcctID;
