--- src/components/AccountByUserID.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { AccountByUserIDConst} from './AccountByUserIDApi';
-
-function AccountByUserID() {
-  return (
-  <div>
-      <AccountByUserIDConst/>
-  </div>
-  )
-}
-
-export default AccountByUserID;
