+++ src/components/account/AccountByUserID.jsx
@@ -1,20 +1,15 @@
 import axios from 'axios';
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 
-export const AccountByUserIDConst =()=>{
-
+export const AccountByUserIDConst = () => {
     const [account, setAccount] = useState([]);
 
-    const fetchAccount=()=>{
-        var userID = 1;
+    useEffect(()=>{
+        var userID = 2;
         axios.get(`http://localhost:8081/accounts/${userID}/all-accounts`).then(res =>{
             console.log(res);
             setAccount(res.data);
-        });
-    }
-
-    useEffect(()=>{
-        fetchAccount();
+        });   
     },[]);
 
     return (
@@ -22,11 +17,9 @@ export const AccountByUserIDConst =()=>{
         {account.map(({accountID, accountType, accountBalance}, index) =>{
             return (
                 <div key={index}>
-                    
                     <h3>Account: {accountID}</h3>
                     <p>Type: {accountType}</p>
                     <p>Balance: {accountBalance}</p>
-                
                 </div>
             )
         })}
