+++ src/components/transaction/CreateSingleTransaction.jsx
@@ -1,12 +1,8 @@
 import { tab } from "@testing-library/user-event/dist/tab";
 import axios from "axios";
-import { useState, useEffect } from "react";
-
-
+import { useState } from "react";
 
 export default function CreateNewTransaction(props) {
-    const transactionsAPI = 'http://localhost:8081/transactions'
-
     const [transactionType, setTransactionType] = useState('');
     const [transactionNote, setTransactionNote] = useState('');
     const [amount, setAmount] = useState(0);
@@ -15,17 +11,16 @@ export default function CreateNewTransaction(props) {
     let newDate = new Date();
     let month = newDate.getMonth() + 1;
     let today = `${month < 10 ? `0${month}` : `${month}`}/${newDate.getDate()}/${newDate.getFullYear()}`;
-  
-    //get userID and accountID from useContext
-        function changeTheValue(e){
-            setTransactionType(e);
-            setTransactionBtn(!transactionBtn);
-        }
 
+    //get userID and accountID from useContext
+    function changeTheValue(e) {
+        setTransactionType(e);
+        setTransactionBtn(!transactionBtn);
+    }
 
     function createNewTransaction() {
         console.log(props.amount);
-        axios.post(transactionsAPI, [{
+        axios.post("http://localhost:8081/transactions", [{
             accountID: 3,
             userID: 1,
             amount: amount,
@@ -33,14 +28,14 @@ export default function CreateNewTransaction(props) {
             transactionNote: transactionNote,
             transactionType: transactionType
         }])
-            .then((response) => {
-                console.log(response.data);
-            })
-            .catch(function (error) {
-                console.log(error);
-            })
-
+        .then((response) => {
+            console.log(response.data);
+        })
+        .catch(function (error) {
+            console.log(error);
+        })
     }
+
     return (
         <>
             <form>
@@ -51,20 +46,12 @@ export default function CreateNewTransaction(props) {
                 <input name="type" type="radio" id="deposit" value="DEPOSIT" onClick={(e) => changeTheValue(e.target.value)} />
                 <label for="deposit">Deposit</label>
 
-
                 <br /><br />
-                Amount<br/>
-                <input
-                    type="number"
-                    value={amount}
-                    onChange={(e) => setAmount(e.target.value)} /><br /><br />
+                Amount<br />
+                <input type="number" value={amount} onChange={(e) => setAmount(e.target.value)} /><br /><br />
 
-                Note<br/>
-                <input
-                    type="text"
-                    value={transactionNote}
-                    onChange={(e) => setTransactionNote(e.target.value)}
-                    placeholder="Note" /><br /><br />
+                Note<br />
+                <input type="text" value={transactionNote} onChange={(e) => setTransactionNote(e.target.value)} placeholder="Note" /><br /><br />
 
                 <button onClick={createNewTransaction}>{transactionBtn ? "WITHDRAW" : "DEPOSIT"}</button>
