+++ src/components/transaction/ViewAllTransactions.jsx
@@ -2,16 +2,10 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 
 export default function ViewAllTransactions() {
-    //instead of showing description, can show green or red according to whether it is deposit or withdraw
-    //list by most recent date
-
-    //i would really like to set this up as a modularization component
-    //pass in any transactions JSON as props and format it as table
-    const transactionsAPI = 'http://localhost:8081/transactions';
     const [transactions, setTransactions] = useState([]);
 
     useEffect(() => {
-        axios.get(transactionsAPI)
+        axios.get("http://localhost:8081/transactions")
             .then((response) => {
                 console.log(response.data);
                 setTransactions(response.data);
